cincinnati: expose dead-end release information via MOTD
After checking whether the current Cincinnati node is a dead-end or
not, write out a new MOTD file at `/run/motd.d/deadend.motd`
reflecting the dead-end status. If the current release is a deadend,
a message is printed in the MOTD containing the reason for the
release being a deadend. An info message is also logged. If not a
dead-end release a blank MOTD is written and no message is logged.

To make the final write atomic, a tempfile is written within the
`/run/motd.d/` directory, and renamed to the location
of the complete MOTD to display.

In this case, "pkexec" utility is used to allow Zincati (which runs as
the `zincati` sysuser) to modify root-owned `/run/motd.d/` in order to
write dead-end release information into a MOTD file, and the polkit rule
will ensure that that utility runs only when its invoked by a `zincati-deadend`
binary and the corresponding user is `zincati`.

Resolves: #90
Co-authored by: Robert Failrley (rob.t.fairley@gmail.com)